requirements.txt
blinker==1.6.2
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
ConfigArgParse==1.7
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.7.0
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.3
msgpack==1.0.5
psutil==5.9.5
pyzmq==25.1.0
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
zope.event==5.0
zope.interface==6.0
python-dotenv==1.0.0
redis~=4.6.0
pymongo~=4.4.1